warning in 64 bits now MIGHT be removed ... - I've only tried in 32 bits ...
git-svn-id: https://barricada.hi.inet/repositorio/samson/trunk@953 4143565c-f3ec-42ea-b729-f8ce0cf5cbc3

+++ apps/samsonCat/main_SamsonCat.cpp
@@ -49,7 +49,7 @@ int main(int argc, const char *argv[])
 	nb = fread(&header, 1, sizeof(ss::KVHeader), file);
 
 	if (nb != sizeof(ss::KVHeader))
-		printf("WARNING: read only %d bytes (wanted to read %d)\n", nb, sizeof(ss::KVHeader));
+		printf("WARNING: read only %d bytes (wanted to read %ld)\n", nb, (long int) sizeof(ss::KVHeader));
 		
 
 	if (!header.check())
